Refactor: Moved Protected file to auth folder for better structure
