+++ src/components/auth/ProtectedRoute.jsx
@@ -1,9 +1,9 @@
 import React, { useEffect, useState } from "react";
 import { Navigate } from "react-router-dom";
 import { useDispatch, useSelector } from "react-redux";
-import { fetchUser } from "../app/features/user/userSlice";
-import { logDebug, logError } from "../utils/logger";
-import ErrorComponent from "./error/ErrorComponent";
+import { fetchUser } from "../../app/features/user/userSlice";
+import { logDebug, logError } from "../../utils/logger";
+import ErrorComponent from "../error/ErrorComponent";
 
 /**
  * Korumalı route bileşeni
